elice-library-main/pyproject.toml
[tool.poetry]
name = "elice-library"
version = "0.1.0"
description = ""
authors = ["yangroro <yang@heechan.kr>"]

[tool.poetry.dependencies]
python = "~3.7"
Flask = "^1.1.2"
SQLAlchemy = "^1.3.23"
PyMySQL = "^1.0.2"
python-dotenv = "^0.15.0"
Flask-SQLAlchemy = "^2.4.4"
bcrypt = "^3.2.0"
Flask-Login = "^0.5.0"
cryptography = "^3.4.5"
email-validator = "^1.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
